Remove receiver connection pooling and restructure
