+++ jms/org.wso2.carbon.transport.jms/src/main/java/org/wso2/carbon/transport/jms/receiver/JMSMessageConsumer.java
@@ -22,7 +22,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.carbon.transport.jms.contract.JMSServerConnectorFuture;
 import org.wso2.carbon.transport.jms.exception.JMSConnectorException;
-import org.wso2.carbon.transport.jms.factory.JMSConnectionFactory;
+import org.wso2.carbon.transport.jms.factory.JMSServerConnectionFactory;
 
 import javax.jms.Connection;
 import javax.jms.Destination;
@@ -41,9 +41,9 @@ public class JMSMessageConsumer implements MessageConsumer {
     private static final Logger logger = LoggerFactory.getLogger(JMSMessageConsumer.class);
 
     /**
-     * The {@link JMSConnectionFactory} instance which is used to create the consumer.
+     * The {@link JMSServerConnectionFactory} instance which is used to create the consumer.
      */
-    private JMSConnectionFactory connectionFactory = null;
+    private JMSServerConnectionFactory connectionFactory = null;
 
     /**
      * The JMS message consumer instance created by the respective JMS client.
@@ -79,16 +79,6 @@ public class JMSMessageConsumer implements MessageConsumer {
      * messages.
      */
     private JMSServerConnectorFuture jmsServerConnectorFuture;
-
-    /**
-     * The {@link String} instance represents the jms connection username.
-     */
-    private String username;
-    /**
-     * The {@link String} instance represents the jms connection password.
-     */
-    private String password;
-
     /**
      * The retry handle which is going to retry connections when failed.
      */
@@ -110,25 +100,21 @@ public class JMSMessageConsumer implements MessageConsumer {
      * @param useReceiver Whether to use consumer.receive or use a listener when consuming messages
      * @param jmsServerConnectorFuture The message processor who is going to process messages consumed from this
      * @param serviceId The service Id which this consumer belongs to
-     * @param username The username to use when connecting to the JMS provider
-     * @param password The password to use when connecting to the JMS provider
      * @param retryInterval The retry interval in milliseconds to retry connection to JMS provider when failed
      * @param maxRetryCount The maximum retry count to retry when connection to the JMS provider fails
      * @throws JMSConnectorException if any error occurred staring consuming data.
      */
-    public JMSMessageConsumer(JMSConnectionFactory connectionFactory, boolean useReceiver,
-                              JMSServerConnectorFuture jmsServerConnectorFuture, String serviceId, String username,
-                              String password, long retryInterval, int maxRetryCount) throws JMSConnectorException {
+    public JMSMessageConsumer(JMSServerConnectionFactory connectionFactory, boolean useReceiver,
+                              JMSServerConnectorFuture jmsServerConnectorFuture, String serviceId,
+            long retryInterval, int maxRetryCount) throws JMSConnectorException {
         this.connectionFactory = connectionFactory;
         this.useReceiver = useReceiver;
-        this.username = username;
-        this.password = password;
         this.jmsServerConnectorFuture = jmsServerConnectorFuture;
         this.serviceId = serviceId;
         this.retryInterval = retryInterval;
         this.maxRetryCount = maxRetryCount;
 
-        retryHandler = new JMSConnectionRetryHandler(this, retryInterval, maxRetryCount);
+        retryHandler = new JMSConnectionRetryHandler(this, this.retryInterval, this.maxRetryCount);
 
         startConsuming();
     }
@@ -139,14 +125,8 @@ public class JMSMessageConsumer implements MessageConsumer {
      * @throws JMSConnectorException if starting consumer fails due to a JMS layer error.
      */
     void startConsuming() throws JMSConnectorException {
-
         try {
-            if (null != username && null != password) {
-                connection = connectionFactory.createConnection(username, password);
-            } else {
-                connection = connectionFactory.createConnection();
-            }
-
+            connection = connectionFactory.createConnection();
             ExceptionListener exceptionListener = connection.getExceptionListener();
             if (exceptionListener == null) {
                 ExceptionListener jmsExceptionListener = new JMSExceptionListener(this);
@@ -188,7 +168,7 @@ public class JMSMessageConsumer implements MessageConsumer {
                 connectionFactory.closeConnection(connection);
                 connection = null;
             }
-        } catch (JMSConnectorException e) {
+        } catch (JMSException e) {
             throw new JMSConnectorException("Error closing connection of JMS Service " + serviceId, e);
         } finally {
             messageConsumer = null;
