+++ jms/org.wso2.carbon.transport.jms/src/main/java/org/wso2/carbon/transport/jms/receiver/JMSMessageConsumerBuilder.java
@@ -20,16 +20,16 @@ package org.wso2.carbon.transport.jms.receiver;
 
 import org.wso2.carbon.transport.jms.contract.JMSServerConnectorFuture;
 import org.wso2.carbon.transport.jms.exception.JMSConnectorException;
-import org.wso2.carbon.transport.jms.factory.JMSConnectionFactory;
+import org.wso2.carbon.transport.jms.factory.JMSServerConnectionFactory;
 
 /**
  * The {@link JMSMessageConsumerBuilder} used to consume messages in carbon jms transport.
  */
 public class JMSMessageConsumerBuilder {
     /**
-     * The {@link JMSConnectionFactory} instance which is used to create the consumer.
+     * The {@link JMSServerConnectionFactory} instance which is used to create the consumer.
      */
-    private JMSConnectionFactory connectionFactory = null;
+    private JMSServerConnectionFactory connectionFactory = null;
 
     /**
      * Tells to use a message receiver instead of a message listener.
@@ -48,15 +48,6 @@ public class JMSMessageConsumerBuilder {
     private JMSServerConnectorFuture jmsServerConnectorFuture;
 
     /**
-     * The {@link String} instance represents the jms connection username.
-     */
-    private String username;
-    /**
-     * The {@link String} instance represents the jms connection password.
-     */
-    private String password;
-
-    /**
      * The retry interval (in milli seconds) if the connection is lost or if the connection cannot be established.
      */
     private long retryInterval = 10000;
@@ -72,7 +63,7 @@ public class JMSMessageConsumerBuilder {
      * @param jmsServerConnectorFuture The message processor who is going to process the consumed messages from this
      * @param serviceId The service Id which invoked this consumer
      */
-    public JMSMessageConsumerBuilder(JMSConnectionFactory connectionFactory,
+    public JMSMessageConsumerBuilder(JMSServerConnectionFactory connectionFactory,
                                      JMSServerConnectorFuture jmsServerConnectorFuture, String serviceId) {
         this.connectionFactory = connectionFactory;
         this.jmsServerConnectorFuture = jmsServerConnectorFuture;
@@ -81,16 +72,6 @@ public class JMSMessageConsumerBuilder {
 
     public JMSMessageConsumerBuilder setUseReceiver(boolean useReceiver) {
         this.useReceiver = useReceiver;
-        return this;
-    }
-
-    public JMSMessageConsumerBuilder setUsername(String username) {
-        this.username = username;
-        return this;
-    }
-
-    public JMSMessageConsumerBuilder setPassword(String password) {
-        this.password = password;
         return this;
     }
 
@@ -111,8 +92,8 @@ public class JMSMessageConsumerBuilder {
      * @throws JMSConnectorException If initializing the consumer fails
      */
     public JMSMessageConsumer build() throws JMSConnectorException {
-        return new JMSMessageConsumer(connectionFactory, useReceiver, jmsServerConnectorFuture, serviceId, username,
-                password, retryInterval, maxRetryCount);
+        return new JMSMessageConsumer(connectionFactory, useReceiver, jmsServerConnectorFuture, serviceId,
+                retryInterval, maxRetryCount);
     }
 
 }
