+++ jms/org.wso2.carbon.transport.jms/src/main/java/org/wso2/carbon/transport/jms/receiver/JMSServerConnectorImpl.java
@@ -24,9 +24,7 @@ import org.wso2.carbon.kernel.utils.StringUtils;
 import org.wso2.carbon.transport.jms.contract.JMSServerConnector;
 import org.wso2.carbon.transport.jms.contract.JMSServerConnectorFuture;
 import org.wso2.carbon.transport.jms.exception.JMSConnectorException;
-import org.wso2.carbon.transport.jms.factory.CachedJMSConnectionFactory;
-import org.wso2.carbon.transport.jms.factory.JMSConnectionFactory;
-import org.wso2.carbon.transport.jms.factory.PooledJMSConnectionFactory;
+import org.wso2.carbon.transport.jms.factory.JMSServerConnectionFactory;
 import org.wso2.carbon.transport.jms.utils.JMSConstants;
 
 import java.util.ArrayList;
@@ -45,10 +43,10 @@ public class JMSServerConnectorImpl implements JMSServerConnector {
     private JMSServerConnectorFuture jmsServerConnectorFuture;
 
     /**
-     * The {@link JMSConnectionFactory} instance represents the jms connection factory related with this server
+     * The {@link JMSServerConnectionFactory} instance represents the jms connection factory related with this server
      * connector.
      */
-    private JMSConnectionFactory jmsConnectionFactory = null;
+    private JMSServerConnectionFactory jmsConnectionFactory = null;
 
     /**
      * The number of concurrent consumers that needs to be created.
@@ -59,14 +57,6 @@ public class JMSServerConnectorImpl implements JMSServerConnector {
      * List of {@link JMSMessageConsumer} instances that are created for this connector instance.
      */
     private List<JMSMessageConsumer> messageConsumers;
-    /**
-     * The {@link String} instance represents the jms connection user-name.
-     */
-    private String userName;
-    /**
-     * The {@link String} instance represents the jms connection password.
-     */
-    private String password;
     /**
      * The {@link Properties} instance represents the jms connection properties.
      */
@@ -101,9 +91,6 @@ public class JMSServerConnectorImpl implements JMSServerConnector {
     private void init(Map<String, String> connectorConfig) throws JMSConnectorException {
         properties = new Properties();
         properties.putAll(connectorConfig);
-
-        userName = connectorConfig.get(JMSConstants.CONNECTION_USERNAME);
-        password = connectorConfig.get(JMSConstants.CONNECTION_PASSWORD);
 
         String retryIntervalParam = connectorConfig.get(JMSConstants.RETRY_INTERVAL);
         if (retryIntervalParam != null) {
@@ -206,24 +193,12 @@ public class JMSServerConnectorImpl implements JMSServerConnector {
     void startConsuming() throws JMSConnectorException {
         try {
             if (jmsConnectionFactory == null) {
-                switch (connectionFactoryNature) {
-                    case JMSConstants.CACHED_CONNECTION_FACTORY :
-                        jmsConnectionFactory = new CachedJMSConnectionFactory(properties);
-                        break;
-                    case JMSConstants.POOLED_CONNECTION_FACTORY :
-                        jmsConnectionFactory = new PooledJMSConnectionFactory(properties);
-                        break;
-                    default :
-                        jmsConnectionFactory = new JMSConnectionFactory(properties);
-                }
+                jmsConnectionFactory = new JMSServerConnectionFactory(properties);
             }
-
             messageConsumers = new ArrayList<>();
             JMSMessageConsumerBuilder consumerBuilder = new JMSMessageConsumerBuilder(jmsConnectionFactory,
                     jmsServerConnectorFuture, serviceId);
             consumerBuilder.setUseReceiver(useReceiver)
-                    .setUsername(userName)
-                    .setPassword(password)
                     .setRetryInterval(retryInterval)
                     .setMaxRetryCount(maxRetryCount);
             for (int i = 0; i < numOfConcurrentConsumers; i++) {
